requirements.txt
Flask==2.1.0
line-bot-sdk==1.20.0